Use uint64 consistently
Previously, we were using byte and uint64 inconsistently.
This required downcasting uint64 to byte, which can cause overflow problems
if we aren't really careful.  This commit switches everything to uint64,
except for fields in the leaf node and the UTF8 tables, where we care
about fitting as much data as possible into the CPU cache.