Tokenize UTF-8 at non-combining character boundaries
In order to display UTF-8 text in a terminal, we need to know
which characters to write to each cell.  Some UTF-8 characters
are "combining" characters that have zero width.  This includes
accent marks and the "zero width joiner" used to build an emoji
from multiple characters.

To handle these cases, tokenize UTF-8 characters by splitting
at non-combining character boundaries.  The display code can
then output each token to a cell, using the first character as
the main character and subsequent characters as combining chars.
(The one exception is an erroneous sequence of zero-width characters,
which the display code can detect by checking the width of the first
character.)