Locator changes to support deletions
Update CharInLine locator to optionally include the newline
or end-of-file position, which will simplify the delete mutation logic.

Add a new locator for returning the cursor onto a line.
This is useful for "fixing" the cursor if it ends up on a newline
or past the end of the text.