Implement mutators for deletion
CompositeMutator executes a sequence of mutators, which is necessary
to express "delete a character then move the cursor onto a line if necessary".

DeleteMutator deletes characters from the cursor up to (but not including)
a position returned by a locator.  This can be used to implement any
deletion (to next character in line, end of line, end of file, previous word, etc)
by plugging in a different locator.