Refactor BreakIter into SegmentIter
Previously, BreakIter would return the offsets of breakpoints between
grapheme clusters.  But most (all?) callers needed to inspect the runes
in the grpaheme clusters.  They were doing this by cloning the rune
iterator and advancing it to keep up with the break iter, resulting
in duplicate work.

This commit refactors BreakIter into SegmentIter, so that the runes
in each segment are always returned to the caller.