Implement scrolling
Scroll the view origin to make the cursor visible,
taking into account line wrapping.

The scrolling algorithm is worst-case O(n) where n is
the number of runes in the text.  The worst-case occurs
in a document with a single line when scrolling at the end
of the text.  I don't think we can do better than this
without caching the line-wrap positions, which would
make things a lot more complicated and brittle.
Fortunately, the worst case is unlikely to happen
in most documents, and even when it does performance
is comparable to vim.