Replace ReaderAtLine with LineStartPosition
For moving a cursor to the start of a line, we need to know
the character position at the start of the line.  Previously,
the text.Tree method would return a reader at the start of the line,
which doesn't provide that information, so replace it with a method
to return the position.