Fix navigating into line with a tab character
Tabs occupy multiple cells, which can affect navigation up/down a line.

For example:

    1: abcd
    2: ef[ tab ]g
    3: hijkl

Navigating from the last character in line 3 up to line 2 should
land the cursor at the start of the tab.  Navigating up again
should land the cursor at the end fo line 1.

Previously, this wouldn't happen because we were counting the number
of grapheme clusters from the start of the line, not the number of cells.
This commit updates the relative line locator to count the number of cells.