Remove view repositioning in scroll up/down commands
The initial attempt to reposition the view when scrolling
up/down using Ctrl-U/Ctrl-D had some ugly behavior
at the end of a document (some lines weren't visible
when scrolling to the very end and then back again).

Going to revert that attempt for now and reimplement
it in a way that's closer to vim (moving the view up/down
by a set number of lines).