Refactor scroll up/down to use more efficient lookup
Previously, we were simulating moving the cursor up N lines,
then moving the cursor backward to the start of the line.
It's cheaper to move the cursor directly to the start of the target
line by looking up the start of the line in the text tree,
so do that instead.