Separate goroutine for polling terminal events
Previously, we were calling a blocking function in tcell
to poll for terminal events within the main event loop.
Eventually, we're going to need to respond to API calls from
the Python REPL as well.  This commit prepares for that change
by moving the blocking call to a separate goroutine.