Simplify Python REPL restart
Previously, we were restarting the REPL in the goroutine that polls output.
This is a potential race condition with undefined behavior, since the goroutine
was modifying e.repl while the main event loop was reading e.repl.

Fix it by having the main event loop poll the REPL output chan directly.
Closing the chan indicates that the REPL needs to be restarted.
Since this is checked in the main event loop, there is no race condition.