Automatically reload file if changes detected
For now, we always reload the file even if there are unsaved changes.
Eventually, the editor can check if there are unsaved changes
before applying the mutator to reload the document.

Since we're implementing loading as a mutator, we could also
load a document from the REPL, but this isn't implemented yet.