DFA matcher uses unicode codepoint positions
The DFA matcher needs to keep track of the match length
and end position in unicode codepoints so we can translate
it back to positions in the document.

Additionally, the DFA matcher needs to be able to check
whether it's reached the end of the document so it can
match on start/end of text.