Fix exponential memory usage of token tree
Previously, we were sometimes materializing all nodes in the full binary tree.
Since the number of nodes is exponential in the depth of the binary tree,
this could cause OOM crashes when tokens were inserted in the rightmost
leaves of the tree.

Fix it by storing binary tree nodes in a hash map, so we materialize
only the nodes that are actually in the tree.