Fix bug in lookahead calculation for empty tokens
Previously, lookahead for empty tokens was set to one past
the end position of the token.  This doesn't work when the tokenizer
looked further ahead without finding a match.

An example where this fails is string matching.  The empty token
might include just the first quote, but a quote added later in the
text could complete the string, so the empty token's lookahead
needs to extend far enough to include the later quote.

To fix it, set the lookahead of the empty token to the max
lookahead from all failed searches by the tokenizer.
Additionally, we need to consider a token affected if the change
occurs at the lookahead position.