Re-use buffer in automata
The temporary buffer used by the DFA in MatchLongest()
was being allocated on the heap (likely because it was passed
as a slice to Read()).  This caused significant overhead to allocate
and GC the buffer.  To avoid this overhead, allocate a buffer
once for the DFA and re-use it for all calls to MatchLongest().

This change results in a ~24% decrease in the time to tokenize JSON.