Fix lookahead length calculation in tokenizer
Previously, entries were using the total lookahead length rather
than the max lookahead length.  In some cases, this would cause the search
for the first affected token to traverse to a leaf node that didn't
have an affected token.  A null iterator would be returned, and the tokenizer
would retokenize from the start of the document.

Fix it by calculating max lookahead length for each entry.