Fix scrolling bug when document ends on newline
The scrolling algorithm was not correctly handling the case
when a cursor is at the end of a document that ends with a newline.
The display code shows the cursor on the following, empty, line;
however, the scroll code was treating the cursor as if it were
at the end of the previous line.