Require path argument
Require the path argument so the editor always knows where
to save the file, even for new files.

If the file does not yet exist, create a placeholder file watcher
with timestamp set to the Unix epoch and an empty checksum.
This allows us to remove the "empty" watcher implementation.