Move split escape key into editor
Previously, the input interpreter was responsible for "splitting"
key events with modifiers into an escape keypress followed by an unmodified
keypress.  This makes it difficult to move the input mode into the editor
state, which is necessary so that menu actions can open new menus
(with input mode "menu").  Fix it by making the editor responsible
for splitting the escape key events.