Reduce number of syscalls in file.Walk
filepath.Walk calls stat for every file, but we don't need
all the extra info from the stat call.  It's faster to
retrieve the file types using os.Readdir.

See https://github.com/golang/go/issues/16399

That issue points to a faster implementation used by go tools
that's multi-threaded.  That may be something to explore
in the future, but it will make the code more complicated
so I'm avoiding it for now.