Avoid creating temporary file
Previously, invoking aretext without a path argument would
always create a temporary file with prefix "untitled-".

This had a few disadvantages:
    1) A temporary file was created even if the user didn't save the document.
    2) The temporary file was created in /tmp, which might be difficult to find later.

A better approach is to choose the path as "untitled-{unique ID}" in the current
working directory.  The file won't be created unless the user saves it
(same behavior as any path that doesn't yet exist).

Choose the unique ID as a unix timestamp.  This will almost always be unique,
and it avoids making extra syscalls to check the filesystem.  (In the worst
case, if the file already exists, it will be loaded like any other file).