Define config and ruleset
Define configuration settings in a struct, which can be built
from applying configuration rules that match the path of a file.

This provides a flexible way to combine rules.  The rules engine
can be agnostic about languages, projects, and other dimensions a
user might define -- it just needs to check if the rule pattern
matches the file path.