Load-or-create config file on aretext startup
Load the config file from $HOME/.config/aretext/config.json on startup,
creating a default config if none exists.

If the config is invalid (e.g. not JSON), exit with an error.

Add a CLI flag "-noconfig" so users can disable config loading.
This is necessary so users can open an invalid config file in aretext.