Codegen the config implementation
To support boolean fields, we need to parse JSON into a struct
with a pointer to a bool.  This allows us to distinguish a null value
(meaning, don't override the config) with a non-null value.

Unfortunately, converting between concrete values and pointers introduces
a lot of boilerplate code, which will be hard to maintain.
Address this by generating everything from a JSON spec that defines
the config fields and validation rules.