Implement custom menu commands
Add configuration for user-defined menu items that execute
a command in a shell.

To avoid complicating the event loop, pipe the output
of the command to a pager (usually `less`).  The pager will
take control of the terminal, allowing the user to scroll and
perform other operations.  Once the command completes and the user
exits the pager, the editor will redraw the screen.

The editor reads the pager and shell from environment variables
($PAGER and $SHELL, respectively) with sensible defaults
(`sh` for the shell, `less -R` for the pager).

This supports use cases like running a formatter, tests, or build
commands.  Future commits will add support for "fire-and-forget"
commands like "run this thing in a new tmux window".