Support unicode character classes in regexp parser
Golang syntax highlighting defines tokens in terms of unicode
character classes, so we need a way to express those in
syntax rule regex.

To produce the NFA, take the union of all runes in the char class,
where each rune is an NFA that concatenates the bytes from
the rune's UTF-8 encoding.  I'm not sure if this is going to be
super slow to compile, but it's simple to implement and trivially
correct, so I'm going to try it.