Update DFA to handle truncated text
Previously, the DFA expected textLen to exactly match the number
of runes output from the reader.  For sub-tokenizers, we want
to run a DFA on a slice of the full text (tokenizing the token
produced by a parent tokenizer).  To make this easier, update
the DFA to process at most textLen runes from the reader,
even if the reader produces more bytes.