Implement subtokenizers
Previously, the syntax parser ran a single tokenizer to segment
the text into tokens.  In some cases, we want to further segment
tokens based on some pattern.  For example:

1) Parsing the start and end quotes of a string as separate tokens
   so they are treated as word boundaries during cursor movement.
2) Applying syntax highlighting in a different programming language
   for code snippets in markdown documents.
3) Highlighting escape characters within a string token.

To support these cases, this commit allows tokenizers to specify
"sub rules" that are used to further tokenize the output of the
parent tokenizer.

Worst-case time complexity is O(nm) where n is the number
of bytes in the full document, and m is the depth of the subparsers.
For example, a rule to tokenize a string might need to process
all the bytes in the document, and a sub-rule to tokenize the start/end
quotes of the string would need to reprocess those bytes.
However, since most tokens are short, this should provide acceptable
performance in practice.