Refactor normal mode input parsing
Previously, the input interpreter used a regex on a buffered sequence
of input runes to recognize normal mode input sequences.  This approach
had several limitations:

1) It could not recognize special characters (e.g. CtrlU), so we handled
   those in a separate code path.
2) It could not handle input sequences where the suffix was a valid command.
   For example "daw" would get interpreted as "a" when the user typed
   the second character.
3) It was difficult to keep the regex in sync with the switch statement.

To address these issues, this commit introduces an input parser
configured with a set of rules.  The parser maintains a set of candidate
rules that could accept the current input sequence; if any candidate
accepts, the parser accepts/resets, and if all candidates reject, the parser
rejects/resets.