Do not require document to exist on reload
When aretext starts without a path argument, it sets the
path to "untitled-{timestamp}.txt" and opens an empty document.
Modifying the document sets the "changed" flag, which prevents
the user from opening another document.  The user cannot discard
changes by reloading, because the file does not exist on disk,
so the reload fails with a file not found error.  This forces
the user to either save the document or quit before opening
another document.

To address this, set the "require exists" option to false
when reloading a document.  If the document doesn't exist,
this will load an empty document.

This seems like reasonable behavior even if the document
previously existed at that path and moved.  Reloading will
accurately reflect the state of the document (empty) and
is consistent with the behavior of starting aretext
with a path to a nonexistent file.