Word movement should skip whitespace around syntax token
Previously, word movements would land on the start of the next
syntax token, even if that token began with whitespace.
For example, pressing "w" from the beginning of a comment line
like "// foobar" would stop at the space, since that's the beginning
of the comment contents token.

Fix it by falling back to the whitespace-based word boundary
if the syntax token boundary lands on a whitespace.