Stop splitting key events with modifiers
In previous versions of tcell, the escape key would sometimes
be interpreted as a modifier on the following key.  The editor
would compensate for this by splitting key events with modifiers
into an escape key event followed by an unmodified key event.

It appears this issue has been fixed in tcell v2.  I removed the
key splitting code and couldn't reproduce the issue I was seeing
before.