Use go:embed to create default config
Previously, aretext would serialize an in-memory config ruleset
to create the default config.  This could lead to subtle bugs
where the in-memory representation differed from the serialized
representation (for example, the in-memory representation might
have a slice of strings, but the deserialized representation would
convert this to a slice of interface{}).

We also eventually want to switch the config format to TOML
and add comments, which would be difficult to do with the in-memory
representation.

To fix these issues, embed the default config directly in the binary.
This allows us to control exactly what the config file contains
and eliminates issues with roundtrip serialization.