Fix savepoint invalidation in undo log
Previously, tracking a new operation with a savepoint in the
future would reset the savepoint to the beginning of the log.
This isn't correct, because the beginning of the log no longer
reflects the last savepoint.  Fix it by setting the savepoint
to -1 so it won't match any location in the log.