Use reverse dim style for selected region
It is difficult to find a style that allows both the cursor
and selection to be visible for all cursor styles (block, line, etc.)

Previously we were using reverse for the selection, except for the cursor
position, which was underlined (to avoid conflict with the reversed style
of the "block" cursor in alacritty).  This looked odd for tabs, since
the first cell would appear reversed+underlined (from the cursor style)
but the other cells in the tab would appear just underlined.

I think reverse dim is the best option, since the cursor should always
be visible, but we can still see the selection.