Shell command no longer uses pager
Previously, a custom menu item would invoke a shell command
with input/output piped to a pager (usually `less`).

This was necessary in older versions of tcell that did not support
suspending/resuming the screen, but now that we're on tcell v2,
we can attach the subshell directly to stdin/stdout/stderr.
This allows us to TUI applications from a custom menu item.
Users that want to page the results can add "| less" to the command
in the menu item configuration.

Also add a success status message so the user can tell that
the command completed.