Fix scroll bug on document reload
When a document was reloaded, the view origin was not reset
before scrolling to the cursor.  When the old view origin
no longer aligned with a line break, the view could
clip the first characters from the first visible line.

Fix it by setting the text origin to 0 before scrolling
to the cursor position on document reload.

Resolves https://github.com/aretext/aretext/issues/38