Add alias for menu items
Allow some search queries to exactly match a menu item.
This eliminates the risk of breaking someone's muscle memory
when new menu items are added, as well as ensuring that common
menu commands are consistent with vim's "ex" mode.

This commit also renames "save" to "write" and assigns it the alias "w"
for consistency with vim.