Use substring search for menu items
Replace the menu search algorithm with a new algorithm
based on substring search.  Items that do not contain
the search query are eliminated, and items that do contain
the search query are ranked based on where the query occurs
(start of string, after punctuation, and after whitespace
are ranked higher).

This is slightly slower than the previous algorithm, but
it produces better results.  For example, searching for
"syntax.go" used to score "syntax/a.go" the same as
"syntax/syntax.go"; now only "syntax/syntax.go" would match.

Resolves https://github.com/aretext/aretext/issues/40