Revert "Menu item clears selection after execution, not before"
This reverts commit be9419cf20bd924c413d705cd1659429e25098a0.

Commit introduced a bug with the "find and open" menu command.
The command sets the input mode to "menu", but the change in
this commit sets it back to "normal", so the user can't
select any items in the menu.

Reverting this until I can figure out a better fix.