Insert shell command output into document
Add new output type "documentInsert" that inserts the
command output into the document at the cursor position,
deleting any existing selection.

Use cases for this include paste from a shared clipboard,
inserting snippets from files, and commenting/uncommenting
a selection.

The shell command output is stored on a dedicated clipboard
page.  Currently, there is no way for the user to access
this page, but it will be once the paste commands allow
the user to specify a register.