Fix incorrect visual mode selection on last empty line
The selected region was being incorrectly clipped when
selecting a last empty line.  In this case, the cursor position
is equal to the document length, but the clipping logic was
selecting the last character, which was a line feed.

Fix it by clipping the positions to the document length,
even if this is an empty selection.

Resolves https://github.com/aretext/aretext/issues/52