Refactor command output to command mode
Once we add support for piping the document into the command stdin,
there will be twice as many input/ouput configurations, but only
some of them make sense.  In particular, any commands that take
control of the terminal should redirect both stdin and stdout/stderr
(otherwise the user can't control the command).

To keep things manageable, replace the output configuration with
a mode configuration that controls both the input and output handling.
This makes explicit which input/output configs are supported.