Fix display artifacts for CRLF
When drawing a carriage return + line feed, the CR was treated
as the main rune with width zero and the LF was treated as a
combining rune.  For some reason I don't fully understand,
this caused tcell to display drawing artifacts.  I've seen
similar issues in the past when the cell width is zero but
there are combining runes sent to the terminal.

Fix it by never setting any whitespace runes.  These aren't
displayed anyway, so we can replace them with spaces
and just set the style.

Resolves https://github.com/aretext/aretext/issues/56