Set path even if initial document load fails
Previously, if the initial document load failed, the file watcher
path would be set to "".  This prevented the user from reloading
the file (since no watcher was set) and from saving the file
(since "" is not a valid path).

This could happen if the config was invalid or if the user lacked
read permissions on the initial file.

Fix it by setting the filewatcher path even if the initial
document load fails.