Validate configuration on startup
Previously, the merged configuration was validated on document load.
This proved inconvenient because (1) an invalid config wouldn't show
an error until you tried to load a file that triggered the error,
and (2) the status bar did not have enough space to show the full
error or instructions to fix the config.

Change the behavior to validate all config rules on startup.
If the configuration is invalid, print an error to the console
with instructions for how to fix the config.