Fix autoindent for new line above (O)
Previously, the insert new line above command (O) would discard
indentation of the current line and start the new line with no
indentation.

Fix it by inserting the newline after any indentation at the start
of the current line, the moving the cursor after the indentation
on the new line.

Resolves https://github.com/aretext/aretext/issues/57