Fix YAML tokenizer parsing of punctuation
Previously, a string like "version: v0.1.2" would parse
the "1.2" in the version as a number.  This was confusing
and inaccurate.  Address it by extending the "symbol" rule
to include periods and hyphens, and add a subtokenizer to
parse these as punctuation (for word movement).

This still isn't completely accurate, but it's an improvement.