Clear whitespace lines after autoindent
Inserting consecutive newlines with autoindent enabled produces
lines with only indentation.  This usually isn't what the
user wants, and git diff will show these lines as errors.

To fix it, delete lines with only whitespace when the user
presses enter or escape in insert mode.

This is very close to vim's behavior except for a
few small edge cases. For example, when entering
then escaping insert mode on a whitespace line,
vim will keep the whitespace and aretext will delete it.
The vim documentation says that whitespace is deleted
when the user presses escape *after* pressing enter
to create a newline, but that's difficult to replicate
in aretext because input handling and state modification
aren't so tightly coupled.

Resolves https://github.com/aretext/aretext/issues/67