Simplify YAML syntax rules
The YAML tokenizer was incorrectly recognizing numbers and strings
within unquoted string values.  This matches vim's behavior,
but it's incorrect and misleading.  I don't see how we can parse
these constructs correctly without supporting more powerful
grammars, so this commit removes all rules except for comments.
It also parses the comment delimiter as a separate token so
it is treated as a word boundary.