Clear the status message after every action
Previously the status message was cleared on transition
from normal mode to other modes.  This caused some
confusing behavior when displaying buffered input
in the status bar, because the status message would be hidden
then displayed again after a command, making it seem like
a new status message.

Address it by clearing the status on every action
in normal or visual mode.