Implement change-to and change-till commands
The change-to and change-till commands are the same as the
delete commands, followed by a transition to insert mode.

This is slightly inconsistent with vim's behavior in the case
that no matching char is found on the line.  Vim treats that as
a failure and skips the transition to insert mode; aretext transitions
to insert mode at the current cursor position.  (Other commands like
"cw" always transition to insert mode, even on an empty line.)
Matching vim's behavior in this case would behavior would greatly
complicate the code, so I'm accepting the inconsistency for now.