Fix selection when repeating visual mode commands
Previously, the repeat-last-action (".") command would
replay the transition to visual mode, cursor movement,
and command.  This caused a problem when selecting
backwards (anchor position > cursor position).
Since the action was "select up, then run command
on current selection", repeating the action would
run the command on the lines *above* the cursor.
This usually isn't what the user wants.  For
indenting/outdenting, the user usually wants to continue
indenting the same lines; and for toggling case the user
usually wants to "untoggle" the same selection.

To fix it, modify all visual mode selection commands
to use a locator to determine the end of the selection.
This allows the action to be repeated, with the selected
region determined by the *current* cursor location
and the locator.

Resolves https://github.com/aretext/aretext/issues/70