Choose colors that look good in 256-color terminals
When using an 8-color terminal (like TERM=screen), tcell was
mapping the colors to one of the eight available colors,
which looked good.  Terminals that support more colors
(like alacritty) would display the actual color, which
looked terrible.

Update the colors to always use one of the 8 colors available
everywhere.  This ensures that the colors look consistent
across 8-color and 256-color terminals.