Revert "Simplify YAML syntax rules"
This reverts commit d7c10bde7c7b3193675536d03127f1a3cf7bb4d7.

Removing the syntax rules for strings caused hash symbols
in a string literal to be incorrectly recognized as comments.
I do not believe it is possible to accurately parse YAML
using the current parser algorithm.  For now, restore the
original syntax highlighting rules, even though they're
not 100% accurate.  The vim syntax rules for YAML have
similar issues, so this at least restores parity
with vim.