Stop using syntax tokens as word boundaries
Previously, the syntax language would affect word cursor movements.
In practice, this was confusing.  For example, the cursor would stop
at a "/" in a Go expression, but not at a "/" in a file path
within a Go string.  Additionally, this approach required additional
rules in the syntax language to parse punctuation and delimiters,
even though these had no effect on syntax highlighting colors.

After this commit, word boundaries are determined solely by whitespace.
A subsequent commit will add word boundaries based on punctuation.

One backwards-incompatible change introduced in this commit
is the removal of the "undefined" syntax language in favor
of the "plaintext" syntax language.  Configs that referenced
"undefined" will need to be updated to use "plaintext" instead.