Implement user-defined macros
Add menu commands to record and replay a user-defined macro.

I chose to implement this using menu commands rather
than normal-mode keybindings because:

1. It's easier to find and more difficult to mistype.
2. It avoids complicating the normal mode input processor
   (interpreting "qa" as start-recording-macro and
    "q" as stop-recording-macro is difficult in the current
   implementation)

A limitation of this approach is that only a single macro
can be recorded at a time.  I find that I almost never
use more than one macro at a time, but if we want to support
that later we can either add more menu commands (record macro "A")
or additional keybindings ("q" and "@" to match vim's keybindings).

This implementation also sets the "repeat last action" to
replay the macro.  This should make it easier to repeat
a macro multiple times with a minimum number of keystrokes.

Finally, there are some changes to set the undo checkpoint
to the start of the replayed macro.  This allows the user
to undo the entire macro in a single "undo" command.

Resolves https://github.com/aretext/aretext/issues/19