Do not allow user-defined macros from visual mode
Menu commands can be executed from visual mode, which means
that user-defined macros could be recorded or replayed
from visual mode. Previously, it was possible to record
a macro in one mode and replay it in another, which could
result in bad states.

In particular, the editor would stop responding to input for
the following sequence:
1) Open a new document and type the contents "a\n\tb\n\tt\c\n\tttd"
2) Enter visual mode.
3) Start recording a macro.
4) Outdent the selection.
5) Stop recording the macro.
6) Exit visual mode.
7) Replay the macro.

To prevent these bad states, disallow macro recording/replay
in visual mode by hiding the menu options.