Add option to save file before running custom menu command
Some custom menu commands (like `go fmt`) read the file contents from disk.
This can cause problems if the file has unsaved changes:

1. `go fmt` changes the file on-disk.
2. The user tries to save the document.
3. The save aborts because the file has changed on-disk.

This leaves the user two choices: either overwrite the changes on-disk
and rerun the fmt command, or reload the changes on-disk and lose
unsaved changes.

To reduce the likelihood of this scenario, add an option to save
the file automatically before running a menu command.