Add file with same path to timeline
Previously, a user could:
1. Open a file foo.txt
2. Run a custom menu command with mode fileLocations
3. Select a different line in the same file foo.txt
4. Select the "open previous document" menu command.

In step (4), the user would see an error that there is no previous
document. This is confusing because the user won't necessarily
remember whether a location was in the same document or a different
document.

Make the behavior consistent by always including the file location
in the timeline, even if reloading the same document.