Fix missing record issue in fuzzy find algorithm
Previously, the algorithm would filter the top 100 items for each
keyword. It is possible that a record could contain all query keywords
but not rank in the top 100 for all of the keywords
(for example, if every record contains the keyword prefix).

Fix it by retrieving all records for each query keyword.
This produced a new bottleneck sorting items in the ranker,
so use a min heap to limit the number of results to 100.

The implementation is noticeably slower for the first few query chars
on 500K records, but still fast enough to use interactively.
I think there are other optimizations that can speed it up,
but will defer those for a subsequent commit.