Use path compression for fuzzy search trie
Previously, the trie used for fuzzy menu search would allocate
a separate node for each byte in a prefix. For randomly generated
strings (such as paths to build artificts), there could be
many unique prefixes in the trie, causing performance issues
when walking trie nodes.

This commit adds path compression to the trie.
Each edge in the trie can now represent multiple characters,
which reduces the number of nodes.

Resolves: https://github.com/aretext/aretext/issues/102