Fix visual mode and repeat last action interactions
Previously, the action to enter visual mode would be recorded
in the "last action" macro. This was a hacky way to ensure
that normal mode would clear the last action macro when
entering visual mode.

This commit addresses the problem in a better way:

1) Stop recording "enter visual mode" in the last action macro.
2) Visual mode actions (indent/outdent, delete, change case, etc.)
   clear the last action macro. This is consistent with the
   behavior in normal mode.

Now when a user enters/exits visual mode, the "next action"
macro preserves whatever the last action was before entering
visual mode.