Replace goreleaser with a shell script
Go releaser doesn't include debug.ReadBuildInfo in the binaries,
because it runs `go build main.go` instead of
`go build github.com/aretex/aretext`.

I think it's possible to get goreleaser to do what I want
(either using proxy builds or waiting for the "skip main check"
option to get released), but since we're not really using
much of the functionality, I decided to replace goreleaser
with a shell script instead.

This reduces the number of dependencies, ensures that
ldflags and debug.ReadBuildInfo are set correctly,
and automates some other steps of the process (checking install docs
version and creating the git tag).