Fix performance issue with delete lines
Previously, lines were deleted one-at-a-time, with each deletion
triggering a syntax highlighting reparse. This is expensive,
especially since we're now coalescing small computations in the
parse tree to reduce memory.

Fix it by deleting all the lines at once, so we just reparse
the document once. This also reduces the number of undo log
entries and avoids the memory allocation to join lines
for the clipboard.