Revert "Coalesce parse computations to reduce memory"
This reverts commit 495e75030003fa06a41a84d61189fbf8c659ea89.

Reverting due to performance regression. Coalescing small
computations into computations of at least N runes reduces
the number of computations in the parse tree, but also
increases the probability that subsequent computations
won't align with the end of a parse func execution.
This can cause pathological cases where most of the
parse tree needs to be recomputed.