Combine small leaves on initial parse
This is another attempt at reducing memory usage.
When loading a large document, the user is likely to edit
only a small part of it. So we can combine small computations
on the initial parse to reduce the height of the parse tree.

The tree will expand as the user edits the document, but
only where the user edits.